Rename to test_{keras, pytorch}.py
