+++ tests/model_selection/test_keras.py
@@ -12,26 +12,16 @@ from sklearn.model_selection import RandomizedSearchCV
 
 from dask_ml.model_selection import IncrementalSearchCV
 
-try:
-    import tensorflow as tf
-    from tensorflow.keras.datasets import mnist as keras_mnist
-    from tensorflow.keras.layers import Dense, Activation, Dropout
-    from tensorflow.keras.models import Sequential
-    from tensorflow.keras.utils import to_categorical
-    from scikeras.wrappers import KerasClassifier, KerasRegressor
-except:
-    pass
+import pytest
+pytest.importorskip("tensorflow")
+pytest.importorskip("scikeras")
 
-try:
-    import torch.optim as optim
-    import torch.nn as nn
-    import torch.nn.functional as F
-    from skorch import NeuralNetClassifier, NeuralNetRegressor
-
-    PYTORCH = True
-except:
-    PYTORCH = False
-    pass
+import tensorflow as tf
+from tensorflow.keras.datasets import mnist as keras_mnist
+from tensorflow.keras.layers import Dense, Activation, Dropout
+from tensorflow.keras.models import Sequential
+from tensorflow.keras.utils import to_categorical
+from scikeras.wrappers import KerasClassifier, KerasRegressor
 
 
 def mnist() -> Tuple[np.ndarray, np.ndarray]:
@@ -80,49 +70,3 @@ def test_keras(c, s, a, b):
     assert search.best_score_ >= 0
 
 
-if PYTORCH:
-
-    class ShallowNet(nn.Module):
-        def __init__(self, n_features=5):
-            super().__init__()
-            self.layer1 = nn.Linear(n_features, 1)
-
-        def forward(self, x):
-            return F.relu(self.layer1(x))
-
-
-@gen_cluster(client=True)
-def test_pytorch(c, s, a, b):
-    pytest.importorskip("torch")
-    pytest.importorskip("skorch")
-
-    n_features = 10
-    defaults = {
-        "callbacks": False,
-        "warm_start": False,
-        "train_split": None,
-        "max_epochs": 1,
-    }
-    model = NeuralNetRegressor(
-        module=ShallowNet,
-        module__n_features=n_features,
-        criterion=nn.MSELoss,
-        optimizer=optim.SGD,
-        optimizer__lr=0.1,
-        batch_size=64,
-        **defaults,
-    )
-
-    model2 = clone(model)
-    assert model.callbacks == False
-    assert model.warm_start == False
-    assert model.train_split is None
-    assert model.max_epochs == 1
-
-    params = {"optimizer__lr": loguniform(1e-3, 1e0)}
-    X, y = make_regression(n_samples=100, n_features=n_features)
-    X = X.astype("float32")
-    y = y.astype("float32").reshape(-1, 1)
-    search = IncrementalSearchCV(model2, params, max_iter=5, decay_rate=None)
-    yield search.fit(X, y)
-    assert search.best_score_ >= 0
